pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "supervertaler"
version = "1.9.186"
description = "Professional AI-enhanced translation workbench with multi-LLM support, glossary system, TM, spellcheck, voice commands, and PyQt6 interface. Batteries included (core)."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "Michael Beijer", email = "info@michaelbeijer.co.uk"},
]
maintainers = [
    {name = "Michael Beijer", email = "info@michaelbeijer.co.uk"},
]
keywords = [
    "translation",
    "CAT",
    "CAT-tool",
    "AI",
    "LLM",
    "GPT",
    "Claude",
    "Gemini",
    "Ollama",
    "glossary",
    "termbase",
    "translation-memory",
    "TM",
    "PyQt6",
    "localization",
    "memoQ",
    "Trados",
    "SDLPPX",
    "XLIFF",
    "voice-commands",
    "spellcheck",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Office/Business",
    "Topic :: Text Processing :: Linguistic",
    "Environment :: X11 Applications :: Qt",
]

dependencies = [
    # Core UI Framework
    "PyQt6>=6.5.0",
    "PyQt6-WebEngine>=6.5.0",
    
    # Document Processing
    "python-docx>=0.8.11",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
    "lxml>=4.9.0",
    
    # AI/LLM Providers (core - lightweight clients)
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "google-generativeai>=0.3.0",
    "requests>=2.28.0",
    
    # Document Conversion & Analysis
    "markitdown>=0.0.1",
    
    # Translation Quality Metrics (Superbench)
    "sacrebleu>=2.3.1",
    
    # Clipboard Operations
    "pyperclip>=1.8.2",

    # Encoding repair (optional in code, included by default)
    "chardet>=5.0.0",
    
    # YAML Support
    "pyyaml>=6.0.0",
    
    # Markdown Rendering
    "markdown>=3.4.0",
    
    # Basic Spellchecking (lightweight, works offline)
    "pyspellchecker>=0.7.0",

    # Voice dictation/commands
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",

    # PDF text extraction
    "PyMuPDF>=1.23.0",

    # MT providers
    "boto3>=1.28.0",
    "deepl>=1.15.0",

    # Advanced spellcheck (Hunspell via pure Python)
    "spylls>=0.1.7",

    # Windows automation / system helpers
    "keyboard>=0.13.5; platform_system == 'Windows'",
    "ahk>=1.0.0; platform_system == 'Windows'",
    "pyautogui>=0.9.54; platform_system == 'Windows'",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
# Local Whisper is intentionally optional because it pulls in heavy PyTorch wheels.
# Voice dictation/commands still work via the OpenAI Whisper API without it.
local-whisper = ["openai-whisper>=20230314"]

# Legacy extras (kept for backward compatibility).
# Some extras are still listed for historical reasons and may be no-ops.
voice = []
web = []
pdf = []
mt = []
hunspell = []
windows = []
core = []
all = []

[project.urls]
Homepage = "https://supervertaler.com"
Repository = "https://github.com/michaelbeijer/Supervertaler.git"
"Bug Tracker" = "https://github.com/michaelbeijer/Supervertaler/issues"
Changelog = "https://github.com/michaelbeijer/Supervertaler/blob/main/CHANGELOG.md"
Documentation = "https://github.com/michaelbeijer/Supervertaler/blob/main/AGENTS.md"
"Author Website" = "https://michaelbeijer.co.uk"

[project.scripts]
supervertaler = "Supervertaler:main"

[tool.setuptools]
py-modules = ["Supervertaler"]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["modules*"]
